overhaul UI to take up less space
feel free to skip this one, it's pretty major. on smaller screens,
all of the borders and margins take up a lot of space, so this is an
attempt to minimize it and maximize readability.

there's a couple other fixes in this PR, but this one commit is entirely
skippable.

also noticed a few other issues and gripes that might be worth filing
tickets for:

- i can't Ctrl-W in the terminal to delete a word, like the terminal
  usually allows you to do.
- thin-screen mode looks broken? it never goes back to the feeds list
- you can scroll down on the content forever, instead of it stopping
  when you reach the end
- HTML markup looks lackluster. i would prefer the following:
  - don't inline links, instead have "link[1]" and at the end have
    "[1]: https://example.com"
  - syntax highlighting, maybe? whether that's italicizing/bolding or
    even making headings colourful, i think it'd help a lot.